@@ -5,6 +5,7 @@
   "main": "dist/bundle.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
+    "postinstall": "npm run build",
     "build": "webpack -p"
   },
   "author": "Autodesk BioNano",
@@ -19,9 +20,10 @@
     "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-jsx-a11y": "^2.2.3",
     "eslint-plugin-react": "^6.8.0",
-    "webpack": "^1.14.0"
+    "webpack": "^2.2.0-rc.3"
   },
   "dependencies": {
-    "keymirror": "^0.1.1"
+    "keymirror": "^0.1.1",
+    "webpack": "^2.2.1"
   }
 }
