@@ -5,6 +5,7 @@ import Status from '../components/status';
 import View from '../components/view';
 import WorkflowRecord from '../records/workflow_record';
 import WorkflowSteps from '../components/workflow_steps';
+import ioUtils from '../utils/io_utils';
 import selectionConstants from '../constants/selection_constants';
 
 require('../../css/workflow.scss');
@@ -14,21 +15,21 @@ function Workflow(props) {
   // TODO this will never happen b/c not displaying nodes anymore
   if (props.selection.type === selectionConstants.WORKFLOW_NODE) {
     const selectedWorkflowNode = props.workflow.workflowNodes.find(
-      workflowNode => workflowNode.id === props.selection.id
+      workflowNode => workflowNode.id === props.selection.id,
     );
     selectedModelData = selectedWorkflowNode.modelData;
   } else if ((props.selection.type === selectionConstants.WORKFLOW_NODE_LOAD ||
     props.selection.type === selectionConstants.WORKFLOW_NODE_EMAIL ||
     props.selection.type === selectionConstants.WORKFLOW_NODE_LIGAND_SELECTION) &&
-    props.workflow.run.inputPdb) {
-    selectedModelData = props.workflow.run.inputPdb;
+    props.workflow.run.inputs.size) {
+    selectedModelData = ioUtils.getInputPdb(props.workflow.run.inputs);
   } else if (props.selection.type ===
     selectionConstants.WORKFLOW_NODE_RESULTS) {
-    if (props.morph === 1) {
-      selectedModelData = props.workflow.run.outputPdb;
-    } else {
-      selectedModelData = props.workflow.run.inputPdb;
-    }
+    const modelDatas = props.workflow.run.inputs
+      .concat(props.workflow.runs.outputs)
+      .map(input => input.pdb);
+
+    selectedModelData = modelDatas.get(props.morph);
   }
 
   let viewError;
@@ -42,7 +43,7 @@ function Workflow(props) {
   let selectionStrings = null;
   if (props.workflow.run.selectedLigand) {
     selectionStrings = props.workflow.run.inputPdbProcessingData.get('ligands').get(
-      props.workflow.run.selectedLigand
+      props.workflow.run.selectedLigand,
     );
   }
 
@@ -61,13 +62,13 @@ function Workflow(props) {
       />
       <Status
         changeLigandSelection={props.changeLigandSelection}
-        fetchingPdb={props.fetchingPdb}
-        fetchingPdbError={props.fetchingPdbError}
+        fetchingPdb={props.workflow.run.fetchingPdb}
+        fetchingPdbError={props.workflow.run.fetchingPdbError}
         morph={props.morph}
         nodes={props.nodes}
         onClickColorize={props.onClickColorize}
         onChangeMorph={props.onChangeMorph}
-        onUpload={props.onUpload}
+        onSelectInputFile={props.onSelectInputFile}
         selectedLigand={props.workflow.run.selectedLigand}
         selection={props.selection}
         submitPdbId={props.submitPdbId}
@@ -85,8 +86,12 @@ function Workflow(props) {
   );
 }
 
+Workflow.defaultProps = {
+  workflow: null,
+};
+
 Workflow.propTypes = {
-  changeLigandSelection: React.PropTypes.func,
+  changeLigandSelection: React.PropTypes.func.isRequired,
   clickAbout: React.PropTypes.func.isRequired,
   clickRun: React.PropTypes.func.isRequired,
   clickWorkflowNodeLigandSelection: React.PropTypes.func.isRequired,
@@ -94,14 +99,12 @@ Workflow.propTypes = {
   clickWorkflowNodeEmail: React.PropTypes.func.isRequired,
   clickWorkflowNodeResults: React.PropTypes.func.isRequired,
   colorized: React.PropTypes.bool.isRequired,
-  fetchingPdb: React.PropTypes.bool,
-  fetchingPdbError: React.PropTypes.string,
   morph: React.PropTypes.number.isRequired,
-  nodes: React.PropTypes.instanceOf(IMap),
+  nodes: React.PropTypes.instanceOf(IMap).isRequired,
   onClickColorize: React.PropTypes.func.isRequired,
   onChangeMorph: React.PropTypes.func.isRequired,
-  onUpload: React.PropTypes.func.isRequired,
-  runPage: React.PropTypes.bool,
+  onSelectInputFile: React.PropTypes.func.isRequired,
+  runPage: React.PropTypes.bool.isRequired,
   selection: React.PropTypes.instanceOf(SelectionRecord).isRequired,
   submitPdbId: React.PropTypes.func.isRequired,
   submitEmail: React.PropTypes.func.isRequired,
