@@ -12,9 +12,9 @@ const WorkflowRecord = new Record({
   fetchingError: null,
   selectLigands: false,
   run: new RunRecord(),
-  runs: 0,
+  runCount: 0,
   title: '',
-  views: 0,
+  viewCount: 0,
   workflowNodes: new IList(),
 });
 
