@@ -1,5 +1,7 @@
 import isEmail from 'validator/lib/isEmail';
 import { statusConstants } from 'molecular-design-applications-shared';
+import apiUtils from './api_utils';
+import ioUtils from './io_utils';
 
 const workflowUtils = {
   /**
@@ -53,7 +55,7 @@ const workflowUtils = {
   },
 
   isRunnable(run) {
-    if (!run.inputPdbUrl) {
+    if (!ioUtils.getInputPdb(run.inputs)) {
       return false;
     }
     if (!isEmail(run.email)) {
@@ -76,6 +78,41 @@ const workflowUtils = {
       reader.onerror = reject;
       reader.readAsText(file);
     });
+  },
+
+  /**
+   * Using the api, go through the full step0 input processing flow
+   * Calls to this should be surrounded by try/catch!
+   * @param workflowId {String}
+   * @param pdb {String}
+   * @returns {Array}
+   */
+  processInput: async function processInput(workflowId, pdb) {
+    const inputs = await apiUtils.processInputPdb(workflowId, pdb);
+
+    // Find the json status input
+    const jsonInput = inputs.find(input =>
+      input.value.lastIndexOf('.json') === input.value.length - 5,
+    );
+    if (!jsonInput) {
+      throw new Error('Expected a json file in step0 response.');
+    }
+
+    // Get the json data for these inputs
+    jsonInput.fetchedValue = await apiUtils.getIoData(jsonInput.value);
+
+    // Get the processed input pdb
+    const pdbInputIndex = inputs.findIndex(input =>
+      input.value.lastIndexOf('.pdb') === input.value.length - 4,
+    );
+    if (pdbInputIndex === -1) {
+      throw new Error('Expected a pdb file in step0 response.');
+    }
+    inputs[pdbInputIndex].fetchedValue = await apiUtils.getPdb(
+      inputs[pdbInputIndex].value,
+    );
+
+    return inputs;
   },
 };
 
