@@ -6,21 +6,55 @@ const testUtils = require('../test/test_utils');
 
 const router = new express.Router();
 
-router.get('/test', (req, res) => {
-  testUtils.runTest2()
+router.get('/', (req, res) => {
+  testUtils.runTestCCC()
     .then((result) => {
-      res.send({success:result.success});
-    }, (err) => {
-      res.sendStatus(500).send(JSON.stringify({test:'failed', error:JSON.stringify(err)}));
+      res.send({success:true, result});
+    })
+    .catch((err) => {
+      res.status(500).send(JSON.stringify({success:false, error:err}));
     });
 });
 
-router.get('/test-all', (req, res) => {
+router.get('/workflow0', (req, res) => {
+  testUtils.runTestWorkflowVDE()
+    .then((result) => {
+      if (result.success) {
+        res.send(result);
+      } else {
+        res.status(500).send(result);
+      }
+    })
+    .catch((err) => {
+      res.status(500).send(JSON.stringify({success:false, error:err}));
+    });
+});
+
+router.get('/workflow1', (req, res) => {
+  testUtils.runTestWorkflowQMMM()
+    .then((result) => {
+      if (result.success) {
+        res.send(result);
+      } else {
+        res.status(500).send(result);
+      }
+    })
+    .catch((err) => {
+      res.status(500).send(JSON.stringify({success:false, error:err}));
+    });
+});
+
+router.get('/all', (req, res) => {
   testUtils.runAllTests()
     .then((result) => {
-      res.send(result);
-    }, (err) => {
-      res.sendStatus(500).send(JSON.stringify({success:false, error:JSON.stringify(err)}));
+      if (result.success) {
+        res.send(result);
+      } else {
+        res.status(500).send(result);
+      }
+    })
+    .catch((err) => {
+      res.status(500).send(JSON.stringify({success:false, error:err}));
     });
 });
 
