@@ -2,11 +2,25 @@ const workflows = [
   {
     id: '0',
     selectLigands: false,
-    title: 'VDE',
+    title: 'Calculate Vertical Detachment Energy',
+    bgIndex: 0,
+    bgColor: '#3762E9',
+    color: '#F1FF66',
+    comingSoon: false,
+    creatorImage: '/assets/logo1.png',
+    description: 'This is the place to put more info regarding this workflow',
+    viewCount: 0,
   }, {
     id: '1',
     selectLigands: true,
     title: 'Preparing the outer ligand structure',
+    bgColor: '#292E60',
+    bgIndex: 1,
+    color: '#2FE695',
+    comingSoon: false,
+    creatorImage: '/assets/logo2.png',
+    description: 'This is the place to put more info regarding this workflow',
+    viewCount: 0,
   },
 ];
 
